// article-editor/types/utils/jsonPageGenerator/MicroMetaArticlePreparedType.ts
type MicroMetaArticlePreparedType = {
  title: string
  description: string
  publicationDate: string
  modifiedDate: string
  previewImage: string
  slug: string
  path?: string
  authors?: string
  ratingValue?: string
  ratingCount?: string
  reviewCount?: string
  bestRating?: string
  worstRating?: string
}

export default MicroMetaArticlePreparedType